move utils into public, add min workaround
