+++ sdk/communication/communication-phone-numbers/test/public/lro.purchaseAndRelease.spec.ts
@@ -5,12 +5,12 @@ import { isPlaybackMode, Recorder, env } from "@azure/test-utils-recorder";
 import { assert } from "chai";
 import { Context } from "mocha";
 import { PhoneNumbersClient, SearchAvailablePhoneNumbersRequest } from "../../src";
-import { matrix } from "../utils/matrix";
+import { matrix } from "./utils/matrix";
 import {
   canCreateRecordedClientWithToken,
   createRecordedClient,
   createRecordedClientWithToken
-} from "../utils/recordedClient";
+} from "./utils/recordedClient";
 
 matrix([[true, false]], async function(useAad) {
   describe(`PhoneNumbersClient - lro - purchase and release${useAad ? " [AAD]" : ""}`, function() {
