+++ sdk/communication/communication-phone-numbers/test/public/utils/mockHttpClients.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { HttpClient, WebResourceLike, HttpOperationResponse } from "@azure/core-http";
-import { PurchasedPhoneNumber } from "../../src";
+import { PurchasedPhoneNumber } from "../../../src";
 
 export const createMockHttpClient = <T = Record<string, unknown>>(
   status: number = 200,
